Update assets and double instead to float

+++ src/GameLibs/Snake/System/AppleSystem.cpp
@@ -42,7 +42,7 @@ void Snake::System::AppleSystem::modifyApplePos(Arcade::ECS::IEventManager &even
     _positions.clear();
 }
 
-void Snake::System::AppleSystem::run(float deltaTime,
+void Snake::System::AppleSystem::run(double deltaTime,
                 Arcade::ECS::IEventManager &eventManager,
                 Arcade::ECS::IEntityManager &currentEntityManager)
 {
